+++ ndsl/dsl/dace/builder/__init__.py
@@ -1,9 +1,9 @@
-from .get_dace_executable import get_dace_executable
+from .builder import get_dace_executable
 from .optimize import optimize_full_program_sdfg
 from .parse import parse_sdfg
 
 __all__ = [
-    "optimize_full_program_sdfg",
     "get_dace_executable",
+    "optimize_full_program_sdfg",
     "parse_sdfg",
 ]
